# Fixed shuffle sharding consistency when zone-awareness is enabled and the shard size is increased or instances in a new zone are added