--- a/pkg/util/yolo.go
+++ b/pkg/util/yolo.go
@@ -1,0 +1,7 @@
+package util
+
+import "unsafe"
+
+func YoloBuf(s string) []byte {
+	return *((*[]byte)(unsafe.Pointer(&s)))
+}