--- a/pkg/util/yolo_test.go
+++ b/pkg/util/yolo_test.go
@@ -1,0 +1,13 @@
+package util
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestYoloBuf(t *testing.T) {
+	s := YoloBuf("hello world")
+
+	require.Equal(t, []byte("hello world"), s)
+}