**What this PR does**:
While working on #3252 I found an issue in the shuffle sharding implementation. When zone-awareness is enabled and the shard size is increased (or instances in a new zone are added) the "consistency" property is not guaranteed, and the returned shard may drop instances which were present in the previous shard (before increasing the shard size). On the contrary, when we increase the shard size, we do expect that the new shard includes all previous instances + new ones.

The problem is due to the fact that the algorithm consider each zone like a separate subring but we're sharing the same pseudo-random generator to select instances in different zones. If you select +1 instance from the 1st zone, the instances selection in the following zone will be based on a different sequence of numbers, effectively leading to have some previous instances removed from the shard.

**Which issue(s) this PR fixes**:
N/A

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`