@pracucci 2020-10-08
Benchmark shows `ShuffleShard` is a bit slower for small shards (because the cost of initialising the random generator). However there's no significative impact on the cached one (which is what we use on the read path).

## `BenchmarkRing_ShuffleShard`

```
benchmark                                                                              old ns/op     new ns/op     delta
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-12        105606        105433        -0.16%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-12       416311        399358        -4.07%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-12       1403560       1345099       -4.17%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-12        101937        119015        +16.75%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-12       459457        458869        -0.13%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-12       1289417       1257109       -2.51%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-12       111629        107648        -3.57%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-12      432635        415295        -4.01%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-12      1426895       1454336       +1.92%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-12       108008        124367        +15.15%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-12      471323        474511        +0.68%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-12      1333308       1293276       -3.00%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-12      100722        101951        +1.22%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-12     398252        391344        -1.73%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-12     1366036       1316046       -3.66%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-12      100346        119046        +18.64%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-12     452132        459570        +1.65%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-12     1261797       1342749       +6.42%

benchmark                                                                              old allocs     new allocs     delta
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-12        25             25             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-12       27             27             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-12       31             31             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-12        43             51             +18.60%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-12       49             57             +16.33%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-12       55             63             +14.55%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-12       25             25             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-12      27             27             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-12      31             31             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-12       43             51             +18.60%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-12      49             57             +16.33%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-12      55             63             +14.55%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-12      25             25             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-12     27             27             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-12     31             31             +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-12      43             51             +18.60%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-12     49             57             +16.33%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-12     55             63             +14.55%

benchmark                                                                              old bytes     new bytes     delta
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-12        67152         67160         +0.01%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-12       223715        223721        +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-12       834713        834725        +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-12        55856         66872         +19.72%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-12       203952        214986        +5.41%
BenchmarkRing_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-12       642809        653812        +1.71%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-12       67152         67160         +0.01%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-12      223717        223723        +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-12      834714        834724        +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-12       55856         66872         +19.72%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-12      203964        214972        +5.40%
BenchmarkRing_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-12      642794        653815        +1.71%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-12      67152         67160         +0.01%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-12     223713        223723        +0.00%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-12     834707        834702        -0.00%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-12      55856         66872         +19.72%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-12     203970        214967        +5.39%
BenchmarkRing_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-12     642790        653816        +1.72%
```

## `BenchmarkRing_ShuffleShardCached`

```
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                                                                    old ns/op     new ns/op     delta
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-12        208           204           -1.92%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-12       524           508           -3.05%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-12       1414          1422          +0.57%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-12        208           200           -3.85%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-12       582           586           +0.69%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-12       1410          1415          +0.35%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-12       208           198           -4.81%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-12      498           512           +2.81%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-12      1387          1376          -0.79%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-12       208           209           +0.48%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-12      594           582           -2.02%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-12      1438          1416          -1.53%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-12      198           209           +5.56%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-12     499           515           +3.21%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-12     1623          1413          -12.94%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-12      200           199           -0.50%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-12     574           568           -1.05%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-12     1421          1431          +0.70%

benchmark                                                                                    old allocs     new allocs     delta
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-12        0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-12       0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-12       0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-12        0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-12       0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-12       0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-12       0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-12      0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-12      0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-12       0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-12      0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-12      0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-12      0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-12     0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-12     0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-12      0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-12     0              0              +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-12     0              0              +0.00%

benchmark                                                                                    old bytes     new bytes     delta
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-12        0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-12       0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-12       1             1             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-12        0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-12       0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-12       0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-12       0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-12      0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-12      1             1             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-12       0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-12      0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-12      0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-12      0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-12     0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-12     1             1             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-12      0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-12     0             0             +0.00%
BenchmarkRing_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-12     0             0             +0.00%
```